Prevent double save dialog boxes by throttling save click
As it takes about half a second for the save dialog box to appear, its easy to click the save button twice and have a second save dialog box appear after you save the image.

Throttle the save button to only be pressed once a second to prevent this from happening.