Use `editorLineNumber.activeForeground` for highlighted line numbers
When highlighting a line the line number becomes harder to see, making it difficult to refer to when referencing the image.
The prior version of CodeSnap set the line number to bright white, but that doesn't necessarily match the theme of the editor.

Here, we're using VS Code's `editorLineNumber.activeForeground` color to style the line number on highlighted lines. This makes the line number clear while still fitting within the editors color theme.